Refactored classes into MVC folder structure
