--- src/main/java/ithacacollege/comp345/group4/classPlanner/Course.java
@@ -1,4 +1,0 @@
-package ithacacollege.comp345.group4.classPlanner;
-
-public class Course {
-}
