--- src/main/java/ithacacollege/comp345/group4/classPlanner/FacultyAPI.java
@@ -1,4 +1,0 @@
-package ithacacollege.comp345.group4.classPlanner;
-
-public class FacultyAPI {
-}
