--- src/main/java/ithacacollege/comp345/group4/classPlanner/User.java
@@ -1,4 +1,0 @@
-package ithacacollege.comp345.group4.classPlanner;
-
-public class User {
-}
