+++ src/main/java/ithacacollege/comp345/group4/classPlanner/controller/FacultyAPI.java
@@ -1,0 +1,4 @@
+package ithacacollege.comp345.group4.classPlanner.controller;
+
+public class FacultyAPI {
+}
