+++ src/main/java/ithacacollege/comp345/group4/classPlanner/controller/StudentAPI.java
@@ -1,0 +1,4 @@
+package ithacacollege.comp345.group4.classPlanner.controller;
+
+public class StudentAPI {
+}
