+++ src/main/java/ithacacollege/comp345/group4/classPlanner/model/Faculty.java
@@ -1,0 +1,4 @@
+package ithacacollege.comp345.group4.classPlanner.model;
+
+public class Faculty {
+}
