+++ src/main/java/ithacacollege/comp345/group4/classPlanner/ui/FacultyUI.java
@@ -1,0 +1,4 @@
+package ithacacollege.comp345.group4.classPlanner.ui;
+
+public class FacultyUI {
+}
